Added option to set imagePullPolicies for KubeVirt and CDI
This patch allows to set imagePullPolicy for KubeVirt and/or CDI by
setting the appropriate variable(s) in the HCO CR under the Spec.
section: KubeVurtImagePullPolicy and/or CDIImagePullPolicy

Please note that currently the CR(s) will NOT be re-created on
the policy change, thus one needs to edit the HCO CR manifest
file.

Signed-off-by: Lev Veyde <lveyde@redhat.com>