Move unit test package install setup into a script
The packages required for unit testing also need to be
installed for other continous integration systems, and not just
for travis-ci. Let's move the install commands into a shell
script and have it integrated with the command that executes
the unit tests.

This way, running "make test" will install the required packages
and also run the unit tests.