Fix template files.
Due to the macro processing issue, an empty line was inserted as the very first line of CRB files.
This was causing issues with tools that use strict checking of the YAML file syntax.
Specifically it was causing the build using the latest origin-operator-registry image to fail.
This patch fixes the templates.

Signed-off-by: Lev Veyde <lveyde@redhat.com>