ssp: deploy MetricsAggregation CR
Starting version 1.0.6, the SSP operator gained the ability
to deploy the prometheus aggregation rules that kubevirt on openshift
wants to expose (e.g. for telemetry).

In this patch, we bump the vendored version to 1.0.7 (last released at
this point in time), and we make sure to create the new CR.

This patch only wants to add code, not to change any of the existing
behaviour.

Signed-off-by: Francesco Romani <fromani@redhat.com>