kvTemplateBundle (fix): HCO watch all namespaces
In order for the HCO to be able to get objects in other namespaces, we
must be watching all namespaces.

It **MAY** be worth enforcing a name/namespace for the HCO CR to ensure
one and only one can ever be installed (and where we tell it to be
installed).