controller: labeller: disable KVM on CI
We should be able to run on CI even without kvm available.
Note this approach is already been taken, and it is supported,
on both node-labeller and ssp-operator CIs.

It wasn't needed before because the incorrect conditions reporting
(fixed in SSP operator 1.0.11) ignored the status of the node-labeller
and incorrectly registered Available=true Progressing=false.

Signed-off-by: Francesco Romani <fromani@redhat.com>