Enable KVM emulation for periodic cnv tests in OpenShift CI
The canary tests are currently broken because KVM emulation is not
enabled for them.

hack/deploy.sh enables KVM emulation when $HOSTNAME = hco-e2e-aws.
The hostname is actually the test name.

For the periodic jobs, the value for $HOSTNAME is e2e-aws-cnv and
that is why emulation was not turned on for them.

Signed-off-by: Richard Su <rwsu@redhat.com>