Migrate from dep to go modules
the operator-sdk doesn't support anymore dep
since v0.12.0, see:
https://github.com/operator-framework/operator-sdk/blob/master/doc/migration/version-upgrade-guide.md#v012x

Move from dep to go modules to be aligned and
ready for future changes (like upgrading the
controller-runtime in a near future).

This change attempts to be less intrusive as possible
trying to use the same versions required with dep.
Upgrades will be handles in other PRs.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>