Skip OperatorGroup creation with CONTENT_ONLY option
Fix the deployment scripts to avoid creating
the OperatorGroup when the user set CONTENT_ONLY mode.

In CONTENT_ONLY mode the script will configure
the environment to be ready for the user
to simply trigger the deployment from OLM UI.
In that case the OperatorGroup will be created
by OLM console according to the user choices
(one or multiple namespaces and which one)
so it should not be created by the deployment script.

Bug-Url: https://bugzilla.redhat.com/1795227

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>