Allow custom names for relatedImages
Enhance the csv-merger tool to let the user
explicitly pass a custom name for each
of the images in relatedImages.

--related-images-list allow the user to pass a
comma separated list of all the images referred
in the CSV for relatedImages field.
relatedImages field contains a list of images where
for each image we can specify image (the pull url)
an name (a label); currently the name is derived from image
while, for different reasons, it could be convenient
let the user pass a custom name.
With this PR --related-images-list is still a comma separated list
but each element could be eventually (if not the code will keep the
old behaviour for backward compatibility) splitted with image|name
syntax.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>