Include manifest generation in sanity tests
Run manifest generation in sanity tests
to ensure that we can only merge PRs where
the automatically generated manifests that are going
to be merged matches what can be generated with their
source configuration.
In order to do that, we have to explicitly tell to
git diff to ignore lines starting with '    createdAt: '
because it will be always updated on freshly generated
CSV files.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>