Emit a warning event on failed uninstall
Now Virt and CDI operators implements a validating
webhook that can prevent their CRs from being deleted
when we have existing VMs or DVs.
When the user will try to delete HCO CR, HCO will try
to delete Virt and CDI CRs and so it can fail there.
In that case, HCO CR is flagged for deletion and HCO
will periodically retry with an exponential
backoff mechanism.

Adding here the code to emit a warning event
to make it more evident to the user.

Adding an explicit delete request on KubeVirt
resource to check its status.

A better approach will be implemented in a future PR:
we are going to implement a validating webhook also here
that will try to delete virt and CDI resources in dry-run
mode and it will reject the deletion of HCO CR with a
clear error message if virt or cdi CRs cannot be deleted.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>