Avoid failing deletion on missing resources
Avoid failing the whole reconciliation loop on deletion
if just one tof the resources we are going to delete
is already missing.

Now Kubevirt and CDI defines a delete protection mechanism
so we can potentially get in status where we sucesfully
deleted Kubevirt CR while we have to loop only on CDI one
until we have pending DVs.
In that case, avoid failing the whole loop just because
we already deleted kubevirt CR in a previous run.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>