Enforce a structural schema for HCO CRD
https://github.com/kubernetes/enhancements/blob/master/keps/sig-api-machinery/20190425-structural-openapi.md#metadata
defines what's a structural schema for a CRD and currently HCO CRD doens't match.

Try to move back to a fully structural schema.
Please notice that the most relevant change is removing the
namespace restriction (which is explictly forbidden) that
can lead HCO CR to be potentially created in any namespace.

We should eventually consider again about moving to a cluster scoped CR.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>