Detect Upgrade Mode
HCO now find out if it in upgrade mode. This is a preparation for the next phase, where HCO will notify the end-user about doing upgrade and completing it.

On init, HCO add the image Id to the CR's annotations.
On non-init reconciliation, HCO checks if ts image Id match to the image Id in the CR; if not, or if there is no image Id in the CR, HCO is in upgrade mode.
On upgrade mode, when processing is completed and HCO become available, HCO updates the image Id in the CR to the new image Id, and exits upgrade mode.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>