Intercept delete in validating webhook
Previously, when trying to delete the HCO resource while VMs exist in
the cluster, the HCO would stay in Terminating state.  This patch
implements a webhook that intercepts delete requests for the HCO.
When validating a delete request, the webhook will try to remove the
KubeVirt and CDI resources in dry run mode, and reject the request in
case the removal of those resources fail.

Signed-off-by: Yuval Turgeman <yturgema@redhat.com>