Avoid deleting KubeVirt Priority Class
KV Priority Class is needed to start virt-operator
but OLM is currently not able to create it so HCO
is trying to create it ASAP.
With https://github.com/kubevirt/hyperconverged-cluster-operator/pull/669
we are trying to create it as soon as HCO starts.
But, for the same reason, we should not trying to
delete it it when the user deletes HCO CR because
virt-operator should continue running even if HCO
CR got deleted.
In the past we were trying to delete the priority
class but we was failing on that just because we forgot
to set app label on that and componentResourceRemoval is
explictly checking for that before really deleting.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>