Create a PriorityClass for KubeVirt on startup (#669) (#686)
Until now, the PriorityClass for KubeVirt was created during
reconciliation (after creating the CR for the HCO).  This caused the
virt-operator to exit with a FailedCreate status, and the CSV to stay in
a Failed status.
This patch creates the PriorityClass when the HCO starts, allowing the
virt-operator to be installed properly and the CSV's status to change to
Succeeded.

Signed-off-by: Yuval Turgeman <yturgema@redhat.com>