Split operator and webhook into two different pods
Currently we are abusing the pod readiness to signal
to OLM that HCO is not ready for an upgrade.
This has a lot of side effects, one of this is the
validating webhook being not able to receive traffic
when exposed by a pod that is not reporting ready=true.
This can cause a lot of side effects if not deadlocks
when the system reach a status where, for any possible
reason, HCO pod cannot be ready and so HCO pod cannot
validate any further update or delete request on HCO CR.
A proper solution is properly use the readiness probe
only to report the pod readiness and communicate
status to OLM via conditions once OLM will be ready for:
https://github.com/operator-framework/enhancements/blob/master/enhancements/operator-conditions.md
in the meanwhile a quick (but dirty!) solution is to
expose the same hco binary on two distinct pods:
the first one will run only the controller and
the second one (almost always ready) just the validating
webhook one.
A deeper code refactor to produce two distinct binaries
is not worth now because we are going to use OLM operator
conditions soon.

Fixes: https://bugzilla.redhat.com/1889401

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>