Add a mutating webhook to protect hco namespace
Introducing a mutating admission webhooks, deployed by
the OLM, bound to the namespace where HCO is running
to protect it from uneducated deletion the can
potentially made it stuck.
The webhook scope is limited to that specific namespace
by the use of an objectSelector.

In this first step the mutating webhook is simply refusing,
with a clear error message,to admit the deletion of the
namespace if the Hyperconverged CR is still there.
A future enhancement can try to delete the Hyperconverged CR
as a side effect (that's why a mutating webhook with side effects)
and, only when it and all its descendants are correctly
and safely deleted, admit the deletion of the namespace.
This requires other ancillary work to make the deletion safe
avoiding race conditions.
So let's simply refuse to delete the namespace as a first
and safer step.

Test it in CI at the end of upgrade test lanes:
the test will try to delete the namespace when the Hyperconverged CR
is still there looking for a specific error message,
the test then will delete the hyperconverged CR and finally
successfully delete the namespace.

Fixes: https://bugzilla.redhat.com/1881874

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>