Change the FeatureGates implemantation (#1060)
Change the FeatureGates implemantation to be strict and to add documentation to the CR for each feature gate.

The previose implementation used a `map[string]bool` to store the feature gates. That way, it was impossible to add documentation for each feature gate.

This PR changes the implementation so the feature gates are now stored in a struct, as a pointer to boolean (to keep them optional). The result is a documentable feature gates as well as strict naming of the feature gates, because their names are not just astring but a field name.

Since `HotplugVolumes` is not a valid json or yaml tag name (starts with upper case), the gate name was changed to `hotplugVolumes` (starts with lower case)

As part of this change, the HyperConverged CRD contains the documentation for the `hotplugVolumes` feature gate, that can be read directly from the CRD, or by calling `kubectl explain`:
```
$ kubectl explain HyperConverged.spec.featureGates
KIND:     HyperConverged
VERSION:  hco.kubevirt.io/v1beta1

RESOURCE: featureGates <Object>

DESCRIPTION:
     featureGates HyperConvergedFeatureGates contains a list of feature enabler
     flags. Setting a flag to `true` will enable the feature. Setting `false` or
     removing the feature gate, disables the feature.

FIELDS:
   hotplugVolumes	<boolean>
     Allow attaching a data volume to a running VMI; type is boolean; default is
     false
```

In addition, this PR adds feature gates documentation to the `docs/cluster-configuration.md` file with the following usage example:

```yaml
apiVersion: hco.kubevirt.io/v1beta1
kind: HyperConverged
metadata:
  name: kubevirt-hyperconverged
spec:
  infra: {}
  workloads: {}
  featureGates:
    hotplugVolumes: true
```

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

**Release note**:
```release-note
Added CRD documentation to the hotplugVolumes feature gate
Added Feature gates documentation to the cluster-configuration.md document
Change HotplugVolumes to hotplugVolumes
```

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>