Try to fix the covaeralls issue
Since HCO moved to run sanity in github action, the coveralls report is wrong - relatives to
the current branch insted of to the master branch, so the coverage check does not show increasing
or decreasing.

This PR moves to use the shogo82148/actions-goveralls@v1 action in order to solve this issue.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>