Fix the Default Values in the FeatureGates field
PR #1109 added default valuse for several feature gates. In order to do that, it tried to set the default value of the `featureGates` field to an empty object (`{}`), in order to force its existance, and then the flag specific default value shoulf take place.

Although this is working very well by manually editing the CRD, and we get exactlly what we want, this is not possible to set the empty object using an annotation:

``` golang
// +kubebuilder:default={withHostModelCPU: true, withHostPassthroughCPU: false}
FeatureGates *HyperConvergedFeatureGates `json:"featureGates,omitempty"`
```

produces a CRD with default of `null`:
``` yaml
featureGates:
  default: null
```
instead of the required:
``` yaml
featureGates:
  default: {}
```
And that causes to the `featureGates` field to not appear by default as required.

This PR sets the default in two places: both on the `featureGates` field as an object, and on `withHostModelCPU` and the `withHostPassthroughCPU` inner fields as a boolean (no change).

After the change we get the required behavior - the default flags appear if they are missing with their default values, independently - it is possible to set each one of them to other value, but deleting them or the whole `fatureGates` field will cause applying of the default values.

Nahshon Unna-Tsameret <nunnatsa@redhat.com>

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>