Add Live Migration Configuration
* Add a new `LiveMigrationConfig` struct to the `HyperConverged`'s `Spec` field.
* Propagate the live migration configuration to the KubeVirt CR.
* Change the `FeatureGates` field to be a struct instead of a pointer to struct, in order to to get the default feature gate values without write them twice.
* Fix bugs in webhook unit test when we go false positive results - some of the test passed but didn't work; e.g. got error, but not the right one

Set the live migration configurations by modifying the fields in the `liveMigrationConfig` under the `spec` field

Bandwidth limit of each migration, in MiB/s. The format is a number and with the `Mi` suffix, e.g. `64Mi`.

**default**: 64Mi

The migration will be canceled if it has not completed in this time, in seconds per GiB of memory. For example, a virtual
machine instance with 6GiB memory will timeout if it has not completed migration in 4800 seconds. If the Migration Method
is BlockMigration, the size of the migrating disks is included in the calculation. The format is a number.

**default**: 800

Number of migrations running in parallel in the cluster. The format is a number.

**default**: 5

Maximum number of outbound migrations per node. The format is a number.

**default**: 2

The migration will be canceled if memory copy fails to make progress in this time, in seconds. The format is a number.

**default**: 150

```yaml
apiVersion: hco.kubevirt.io/v1beta1
kind: HyperConverged
metadata:
  name: kubevirt-hyperconverged
spec:
  liveMigrationConfig:
    bandwidthPerMigration: 64Mi
    completionTimeoutPerGiB: 800
    parallelMigrationsPerCluster: 5
    parallelOutboundMigrationsPerNode: 2
    progressTimeout: 150
```

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>