Remove KubeVirt ConfigMap
KubeVirt is about to drop the configMap support, and will only configurable by modifiying the KubeVirt custom resource.

This PR removes the KubeVirt confgMap. It does not create the configMap and does not reconcile it.

Upon upgrading from previous version, HCO now removes the KubeVirt configMap if exists. Before deleting the configMap, HCO adopt the relevant configurations into the HyperConverged custome resource. The only field that actually abopted is the `migration` field that is adopted into the HyperConverged CR's `LifeMigrationConfig` field. In addition, HCO copies the configMap to a new one, with the name of "kubevirt-config-backup".

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>